create setup.py
