+++ .pre-commit-config.yaml
@@ -1,0 +1,19 @@
+repos:
+-   repo: https://github.com/python/black
+    rev: 22.3.0
+    hooks:
+      - id: black
+
+-   repo: https://github.com/pycqa/isort
+    rev: 5.10.1
+    hooks:
+      - id: isort
+        args: ["--profile", "black", "--filter-files"]
+
+# -   repo: https://gitlab.com/pycqa/flake8.git
+#     rev: 4.0.1
+#     hooks:
+#       - id: flake8
+#         additional_dependencies:
+#           - flake8-black>=0.1.1
+#         language_version: python3
