+++ setup.py
@@ -1,0 +1,26 @@
+from setuptools import find_packages, setup
+
+
+def read_requirements(path):
+    with open(path, "r") as f:
+        return [line.strip() for line in f if not line.isspace()]
+
+
+with open("README.md", "r", encoding="UTF-8") as fh:
+    long_description = fh.read()
+
+
+setup(
+    name="spaco2",
+    version="0.1",
+    python_requires=">=3.7",
+    install_requires=read_requirements("requirements.txt"),
+    packages=["spaco2"],
+    author="Qianhua Zhu, Zehua Jing",
+    author_email="jingzehua@genomics.cn",
+    description="Spaco2: Spatially constrained colorization for visualizing spatial data at single-cell resolution",
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    license="GPL",
+    url="https://github.com/Bai-Lab/Spaco2",
+)
