+++ spaco2/__init__.py
@@ -1,0 +1,1 @@
+from .scm import scm
